requirements.txt
fastapi==0.136.0
uvicorn==0.45.0
httpx==0.28.1
websockets==16.0
argon2_cffi==25.1.0
aiokafka==0.13.0
SQLAlchemy==2.0.49
pydantic-settings==2.14.0
psycopg2==2.9.12
Jinja2==3.1.6
PyJWT==2.12.1
passlib==1.7.4